requirements.txt
Flask==2.1.2
Werkzeug==2.1.2
SQLAlchemy==1.4.36
PyMySQL==1.0.2
gunicorn==20.1.0
python-dotenv
flask_cors
flask_socketio
cryptography
openai
pandas
reportlab
fpdf
requests
google-api-python-client 
google-auth 
google-auth-oauthlib